requirements.txt
python-docx
python-telegram-bot==20.3
rapidfuzz

